Fix component structure
