+++ packages/commons-ui-next/src/Link/Link.test.js
@@ -1,9 +1,9 @@
-import { render } from "@testing-library/react";
+import { render } from "@commons-ui/testing-library";
 import React from "react";
 
-import Link from ".";
+import Link from "./Link";
 
-describe("Link", () => {
+describe("<Link />", () => {
   it("renders unchanged", () => {
     const { container } = render(<Link href="/" />);
     expect(container).toMatchSnapshot();
